refactor(maestro): split virtual_code/generator into cohesive modules
